refactor: move send flow into components
